docs/api/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
flask==3.1.0
PyJWT==2.6.0
cryptography==38.0.4
psycopg2-binary==2.9.10
requests==2.28.1
pytest==7.2.0
grpcio==1.66.1
grpcio-tools==1.66.1